Move more code into the managed module.
