+++ cli/npm/managed/registry.rs
